# backend/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Tenant Payment & Communication API",
    "description": "REST API for property management, lease handling, payment processing, and tenant communication. Includes user authentication, property management, payment tracking, maintenance requests, messaging, and AI-powered features.",
    "version": "1.0.0",
    "contact": {
      "name": "API Support",
      "email": "support@example.com"
    },
    "license": {
      "name": "MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3001",
      "description": "Development server"
    },
    {
      "url": "https://api.example.com",
      "description": "Production server"
    }
  ],
...
  "paths": {
...
    "/api/auth/register": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Register new user",
        "description": "Create a new user account (landlord or tenant)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email", "password", "firstName", "lastName", "role"],
                "properties": {
                  "email": { "type": "string", "format": "email", "example": "user@example.com" },
                  "password": { "type": "string", "format": "password", "minLength": 8, "example": "Password123!" },
                  "firstName": { "type": "string", "example": "John" },
                  "lastName": { "type": "string", "example": "Doe" },
                  "phone": { "type": "string", "example": "+1234567890" },
                  "role": { "type": "string", "enum": ["landlord", "tenant"], "example": "landlord" }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User registered successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "message": { "type": "string", "example": "Registration successful" },
                    "data": {
                      "type": "object",
                      "properties": {
                        "user": { "$ref": "#/components/schemas/User" },
                        "token": { "type": "string", "description": "JWT access token" },
                        "refreshToken": { "type": "string", "description": "JWT refresh token" }
                      }
                    }
                  }
                }
              }
            }
          },
          "409": {
            "description": "Email already registered",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Login user",
        "description": "Authenticate user and receive JWT tokens",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email", "password"],
                "properties": {
                  "email": { "type": "string", "format": "email", "example": "landlord@example.com" },
                  "password": { "type": "string", "format": "password", "example": "Password123!" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "message": { "type": "string", "example": "Login successful" },
                    "data": {
                      "type": "object",
                      "properties": {
                        "user": { "$ref": "#/components/schemas/User" },
                        "token": { "type": "string", "description": "JWT access token (7 days valid)" },
                        "refreshToken": { "type": "string", "description": "JWT refresh token (30 days valid)" }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
    "/api/auth/me": {
      "get": {
        "tags": ["Authentication"],
        "summary": "Get current user",
        "description": "Get details of the authenticated user",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "User details retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "data": {
                      "type": "object",
                      "properties": {
                        "user": { "$ref": "#/components/schemas/User" }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - token missing or invalid",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
...
    "/api/leases": {
      "get": {
        "tags": ["Leases"],
        "summary": "List leases",
        "description": "Get leases (landlords see all, tenants see their own)",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": { "type": "integer", "default": 1 }
          },
          {
            "name": "status",
            "in": "query",
            "schema": { "type": "string", "enum": ["active", "terminated", "pending"] }
          }
        ],
        "responses": {
          "200": {
            "description": "Leases retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "data": {
                      "type": "object",
                      "properties": {
                        "leases": {
                          "type": "array",
                          "items": { "$ref": "#/components/schemas/Lease" }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Leases"],
        "summary": "Create lease",
        "description": "Create a new lease (landlord only)",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["unitId", "tenantId", "startDate", "endDate", "monthlyRent"],
                "properties": {
                  "unitId": { "type": "string", "format": "uuid" },
                  "tenantId": { "type": "string", "format": "uuid" },
                  "startDate": { "type": "string", "format": "date", "example": "2024-01-01" },
                  "endDate": { "type": "string", "format": "date", "example": "2025-01-01" },
                  "monthlyRent": { "type": "number", "example": 1500.00 },
                  "securityDeposit": { "type": "number", "example": 1500.00 },
                  "paymentDueDay": { "type": "integer", "example": 1 },
                  "lateFeeAmount": { "type": "number", "example": 50.00 }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Lease created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "data": { "$ref": "#/components/schemas/Lease" }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter JWT token received from login endpoint"
      }
    },
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "email": { "type": "string", "format": "email" },
          "firstName": { "type": "string" },
          "lastName": { "type": "string" },
          "phone": { "type": "string" },
          "role": { "type": "string", "enum": ["landlord", "tenant"] },
          "profileImage": { "type": "string", "nullable": true },
          "isActive": { "type": "boolean" },
          "notificationPreferences": {
            "type": "object",
            "properties": {
              "email": { "type": "boolean" },
              "push": { "type": "boolean" },
              "sms": { "type": "boolean" }
            }
          },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
...
      "Lease": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "unitId": { "type": "string", "format": "uuid" },
          "tenantId": { "type": "string", "format": "uuid" },
          "landlordId": { "type": "string", "format": "uuid" },
          "startDate": { "type": "string", "format": "date" },
          "endDate": { "type": "string", "format": "date" },
          "monthlyRent": { "type": "number" },
          "securityDeposit": { "type": "number" },
          "paymentDueDay": { "type": "integer" },
          "lateFeeAmount": { "type": "number" },
          "status": { "type": "string", "enum": ["active", "terminated", "pending"] },
          "createdAt": { "type": "string", "format": "date-time" }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean", "example": false },
          "message": { "type": "string" },
          "code": { "type": "string" }
        }
      }
    }
  }
}